\name{ewmaVol}
\alias{ewmaVol}
\title{ewmaVol}
\description{
Esta función calcula la media móvil exponencial, para n periodos, con factor de decaimiento lambda

}
\usage{
ewmaVol(rets,lambda)
}
\arguments{
\item{rets}{Vector con los valores de la serie de tiempo}
\item{lambda}{valor numérico que define el factor de decaimiento exponencial (por ejemplo, 0.94 es el factor recomendado por RiskMetrics para su método de cálculo de volatilidad)}


}

\examples{
AsApAInegi(serieslist, seriesnames, token_inegi)
}
